--- a/pyproject.toml
+++ b/pyproject.toml
@@ -57,6 +57,9 @@ scripts.em = "em_keyboard:cli"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/em_keyboard/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"